resolved error in switching function for e3b

+++ src/Mod.h
@@ -41,9 +41,9 @@ namespace Mod {
 
     template<bool>
     __global__ void scaleSystem_cu(float4 *xs, int nAtoms, float3 lo, float3 rectLen, float3 scaleBy,int *idToIdxs, 
-                                         int *notRigidBody);
+                                         uint8_t *notRigidBody);
     template<bool>
-    __global__ void scaleSystemGroup_cu(float4 *xs, int nAtoms, float3 lo, float3 rectLen, float3 scaleBy, uint32_t groupTag, float4 *fs, int *idToIdxs, int *notRigidBody);
+    __global__ void scaleSystemGroup_cu(float4 *xs, int nAtoms, float3 lo, float3 rectLen, float3 scaleBy, uint32_t groupTag, float4 *fs, int *idToIdxs, uint8_t *notRigidBody);
     void scaleSystem(State *, float3 scaleBy, uint32_t groupTag=1);
     //__global__ void skewAtomsFromZero(cudaSurfaceObject_t xs, float4 xFinal, float4 yFinal);
     //__global__ void skewAtoms(cudaSurfaceObject_t xs, float4 xOrig, float4 xFinal, float4 yOrig, float4 yFinal);
